Added Exclude_tree options
Added the exclude_tree options to both the update_dbtree() and update_fieldtree() methods in the flt_query class.
Also fixed a bug in the Flight class where the set_database method would not populate the root folder if the fieldtree table was empty. it looks like the only way to populate this before was by using the generate_preset_fieldtree() method updating the fieldtree.
Updated the README file with descriptions of the new goodies